verify all contracts in setup chain contracts, refactor
